refactor: ChatMessages out of ee
