+++ web/src/components/ChatMessages/types.ts
@@ -1,0 +1,13 @@
+import type { ChatMessageRole, ChatMessageWithId } from "@langfuse/shared";
+
+
+export type MessagesContext = {
+  messages: ChatMessageWithId[];
+  addMessage: (role: ChatMessageRole, content?: string) => ChatMessageWithId;
+  deleteMessage: (id: string) => void;
+  updateMessage: <Key extends keyof ChatMessageWithId>(
+    id: string,
+    key: Key,
+    value: ChatMessageWithId[Key]
+  ) => void;
+};
