+++ web/src/components/ChatMessages/utils/createEmptyMessage.ts
